Add spatial color curves to shimmer effect and a new shimmer algorithm with greater control. Old algorithm remains as an option and will default for old sequences.

+++ xLights/RenderBuffer.h
@@ -436,7 +436,7 @@ public:
 
     HSVValue Get2ColorAdditive(HSVValue& hsv1, HSVValue& hsv2);
     float GetEffectTimeIntervalPosition();
-    float GetEffectTimeIntervalPosition(float cycles);
+    double GetEffectTimeIntervalPosition(double cycles);
 
     bool _onlyOnMain;
     PathDrawingContext * GetPathDrawingContext();
